% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_ueq.R
\name{analyze_ueq}
\alias{analyze_ueq}
\title{Summarize results for UEQ data}
\usage{
analyze_ueq(df, is.clean = FALSE, ueq_range = c(1:26),
  group_var = 27)
}
\arguments{
\item{df}{dataframe with 26 columns in the order specified in the UEQ documentation}

\item{is.clean}{Boolean whether the passed dataframe is already cleaned}

\item{ueq_range}{the range of columns where the 26 UEQ Items are to be found}

\item{group_var}{index of additional variables that should not be dropped}
}
\value{
Mean comparisons
}
\description{
update
}
