% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ueq.R
\name{clean_ueq}
\alias{clean_ueq}
\title{Clean an import with UEQ items}
\usage{
clean_ueq(df, ueq_range = c(1:26), add_vars = NULL)
}
\arguments{
\item{df}{dataframe with 26 columns in the order specified in the UEQ documentation}

\item{ueq_range}{the range of columns where the 26 UEQ Items are to be found}

\item{add_vars}{index of additional variables that should not be dropped}
}
\value{
dataframe with 26 columns minus the number of rows where the above pattern was true
}
\description{
Cleans an import of the 26 items of the UEQ after the method specified
at \url{https://www.ueq-online.org}. If the difference between the best
and worst evaluation of an item differs by more than 3 this is seen as
an indicator for a problematic data pattern. Rows where this is the case
for 3 or more scales will be removed.
}
