% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_ueq.R
\name{compare_ueq}
\alias{compare_ueq}
\title{Compare two datasets with UEQ items}
\usage{
compare_ueq(df1, df2, range1 = c(1:26), range2 = c(1:26))
}
\arguments{
\item{df1}{first dataset to compare}

\item{df2}{second dataset to compare}

\item{range1}{range of df1}

\item{range2}{range of df2}
}
\description{
Compare two datasets with UEQ items
}
