% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ueq_case_scale_means.R
\name{ueq_case_scale_means}
\alias{ueq_case_scale_means}
\title{Get scale means per case to do further analysis}
\usage{
ueq_case_scale_means(df, is.clean = FALSE, ueq_range = c(1:26),
  add_vars = NULL)
}
\arguments{
\item{df}{dataframe with 26 columns in the order specified in the UEQ documentation}

\item{is.clean}{Boolean whether the passed dataframe is already cleaned}

\item{ueq_range}{the range of columns where the 26 UEQ Items are to be found}

\item{add_vars}{index of additional variables that should not be dropped}
}
\value{
dataframe with 26 columns minus the number of rows where the above pattern was true
}
\description{
Calculates means for each scale per case. Can be used to further analyse differences
in mean values.
}
