% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ueq_items.R
\name{ueq_items}
\alias{ueq_items}
\title{Summarize results for the 26 UEQ items}
\usage{
ueq_items(df, is.clean = FALSE, ueq_range = c(1:26))
}
\arguments{
\item{df}{dataframe with 26 columns in the order specified in the UEQ documentation}

\item{is.clean}{Boolean whether the passed dataframe is already cleaned}

\item{ueq_range}{the range of columns where the 26 UEQ Items are to be found}
}
\value{
dataframe with 9 columns and one row for each of the 26 items
}
\description{
Creates a dataframe with mean, variance, standard deviation,
N, confidence, left and right confidence interval, scale, and item
}
