% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ueq_scales.R
\name{ueq_scales}
\alias{ueq_scales}
\title{Summarize results for the six UEQ scales}
\usage{
ueq_scales(df, is.clean = FALSE, ueq_range = c(1:26))
}
\arguments{
\item{df}{dataframe with 26 columns in the order specified in the UEQ documentation}

\item{is.clean}{Boolean whether the passed dataframe is already cleaned}

\item{ueq_range}{the range of columns where the 26 UEQ Items are to be found}
}
\value{
dataframe with 11 columns and one row for each of the six scales
}
\description{
Creates a dataframe with scale name, mean, variance, standard deviation,
N, confidence, left and right confidence interval, alpha, lambda^2, and
the benchmark comparison.
}
