% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_items.R
\docType{data}
\name{sample_items}
\alias{sample_items}
\title{UEQ sample data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 255 rows and 28 columns.}
\source{
\href{https://www.ueq-online.org}{UEQ Online}
}
\usage{
data(sample_items)
}
\description{
Data from https://www.ueq-online.org, with the 26 UEQ items on a scale from 1 to 7
}
\keyword{datasets}
